requirements.txt
streamlit
requests

SpeechRecognition
google-generativeai
